Remove unused CSS, split shared CSS into separate files
