+++ packages/web/src/components/tipping/support/TopSupporters.module.css
@@ -1,0 +1,8 @@
+.container {
+  margin-top: 32px;
+}
+
+.trophyIcon {
+  width: 18px;
+  height: 18px;
+}
